Alpha Release
This is the first published version of "Card Matching". A card-matching game where users click on cards to "flip" them and expose the color or image underneath. The objective of the game is to find each pair of matching cards. The objective of the project is to hone my full-stack web development skills in a way that is iterative and fun. Hopefully someone (else) will also get some joy out of the experience.